#23 Updated the ability to completely delete all the data for a configurations

+++ BurnSoft.Applications.MLL/ConfigSheets/ConfigListDataName.cs
@@ -368,7 +368,7 @@ namespace BurnSoft.Applications.MLL.ConfigSheets
             return bAns;
         }
         /// <summary>
-        /// Deletes the specified database path.
+        /// Deletes The Configuration and all the data relating to it.
         /// </summary>
         /// <param name="databasePath">The database path.</param>
         /// <param name="id">The identifier.</param>
@@ -380,6 +380,14 @@ namespace BurnSoft.Applications.MLL.ConfigSheets
             bool bAns = false;
             try
             {
+                if (!ConfigListDataMetalic.DeleteByConfigId(databasePath, id, out errOut);
+                if (errOut.Length > 0) throw new Exception(errOut);
+                if (!ConfigListDataPowder.DeleteByConfigId(databasePath, id, out errOut);
+                if (errOut.Length > 0) throw new Exception(errOut);
+                if (!ConfigListDataPowderShotGun.DeleteByConfigId(databasePath, id, out errOut);
+                if (errOut.Length > 0) throw new Exception(errOut);
+                if (!ConfigListDataShotgun.DeleteByConfigId(databasePath, id, out errOut);
+                if (errOut.Length > 0) throw new Exception(errOut);
                 string sql = $"DELETE from Config_List_Name where id={id}";
                 bAns = Database.Execute(databasePath, sql, out errOut);
             }
